requirements.txt
pillow==11.0.0
django==5.1.3
django-htmx==1.21.0
django-allauth==65.2.0
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.10.0
cryptography==43.0.0